Move subscribe routes into separate controller
This is in preparation for adding more digipack routes
