+++ app/controllers/Application.scala
@@ -58,21 +58,6 @@ class Application(
     Redirect(redirectUrl, request.queryString, status = FOUND)
   }
 
-  def subscribeGeoRedirect: Action[AnyContent] = GeoTargetedCachedAction() { implicit request =>
-    val redirectUrl = request.fastlyCountry match {
-      case Some(UK) => "/uk/subscribe"
-      case _ => "https://subscribe.theguardian.com"
-    }
-
-    Redirect(redirectUrl, request.queryString, status = FOUND)
-  }
-
-  def subscribeRedirect(countryCode: String): Action[AnyContent] = CachedAction() { implicit request =>
-    // Country code is required here because it's a parameter in the route.
-    // But we don't actually use it.
-    Redirect("https://subscribe.theguardian.com", request.queryString, status = FOUND)
-  }
-
   def redirect(location: String): Action[AnyContent] = CachedAction() { implicit request =>
     Redirect(location, request.queryString, status = FOUND)
   }
@@ -106,10 +91,6 @@ class Application(
       css = "contributionsLandingPageStyles.css",
       paymentApiPayPalEndpoint = paymentAPIService.payPalCreatePaymentEndpoint
     ))
-  }
-
-  def subscriptionsLanding(title: String, id: String, js: String): Action[AnyContent] = CachedAction() { implicit request =>
-    Ok(views.html.main(title, id, js, description = Some(stringsConfig.subscriptionsLandingDescription)))
   }
 
   def reactTemplate(title: String, id: String, js: String): Action[AnyContent] = CachedAction() { implicit request =>
