+++ app/wiring/AppComponents.scala
@@ -45,6 +45,7 @@ trait AppComponents extends PlayComponents
     regularContributionsController,
     identityController,
     oneOffContributions,
+    subscriptionsController,
     loginController,
     testUsersController,
     payPalNvpController,
